Add persistence disabled test
